#16
Closes #16, httpclient is now 6 separate source files.
