+++ source/Client/HttpClientBuilder.Delete.cs.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace HttpClientBuilder;
+
+internal sealed partial class HttpBuilderClient
+{
+}
