+++ source/Client/HttpClientBuilder.Post.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace HttpClientBuilder.Client
+{
+    internal class HttpClientBuilder
+    {
+    }
+}
